# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_uniedit_server_internal_module_ai_task.Error:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  internal_module_ai_handler.TaskResponse:
    properties:
      completed_at:
        type: integer
      created_at:
        type: integer
      error:
        $ref: '#/definitions/github_com_uniedit_server_internal_module_ai_task.Error'
      id:
        type: string
      progress:
        type: integer
      status:
        type: string
      type:
        type: string
      updated_at:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@uniedit.io
    name: UniEdit Support
    url: https://uniedit.io/support
  description: UniEdit 视频编辑器后端 API
  license:
    name: Proprietary
    url: https://uniedit.io/license
  termsOfService: https://uniedit.io/terms
  title: UniEdit Server API
  version: "1.0"
paths:
...
  /ai/tasks:
    get:
      consumes:
      - application/json
      description: List all AI tasks for the current user with optional filtering
      parameters:
      - description: Filter by status (pending, processing, completed, failed, cancelled)
        in: query
        name: status
        type: string
      - description: Filter by task type (video_generation, image_generation)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List AI tasks
      tags:
      - AI Tasks
  /ai/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Cancel a pending or processing AI task
      parameters:
      - description: Task ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task cancelled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid task ID or task cannot be cancelled
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel AI task
      tags:
      - AI Tasks
    get:
      consumes:
      - application/json
      description: Get details of a specific AI task by ID
      parameters:
      - description: Task ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_ai_handler.TaskResponse'
        "400":
          description: Invalid task ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get AI task
      tags:
      - AI Tasks
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"